+++ mathics/builtin/atomic/numbers.py
@@ -29,15 +29,15 @@ from mathics.core.atoms import (
     Rational,
     Real,
     SymbolDivide,
-    from_python,
 )
 from mathics.core.attributes import (
     listable,
     protected,
 )
+from mathics.core.convert.expression import to_mathics_list
+from mathics.core.convert.python import from_python
 from mathics.core.evaluators import apply_N
 from mathics.core.expression import Expression
-from mathics.core.convert.expression import to_mathics_list
 from mathics.core.list import ListExpression
 from mathics.core.number import (
     dps,
