+++ mathics/builtin/datentime.py
@@ -16,9 +16,10 @@ from datetime import datetime, timedelta
 import dateutil.parser
 
 from mathics.builtin.base import Builtin, Predefined
-from mathics.core.atoms import Integer, Real, String, from_python
+from mathics.core.atoms import Integer, Real, String
 from mathics.core.attributes import hold_all, no_attributes, protected, read_protected
 from mathics.core.convert.expression import to_expression, to_mathics_list
+from mathics.core.convert.python import from_python
 from mathics.core.evaluation import TimeoutInterrupt, run_with_timeout_and_stack
 from mathics.core.element import ImmutableValueMixin
 from mathics.core.expression import Expression
