+++ mathics/builtin/drawing/plot.py
@@ -25,11 +25,11 @@ from mathics.core.atoms import (
     Integer,
     Integer0,
     Integer1,
-    from_python,
 )
 from mathics.core.attributes import hold_all, protected
-from mathics.core.evaluators import apply_N
 from mathics.core.convert.expression import to_expression, to_mathics_list
+from mathics.core.convert.python import from_python
+from mathics.core.evaluators import apply_N
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
 from mathics.core.symbols import Symbol, SymbolList, SymbolN, SymbolPower, SymbolTrue
