+++ mathics/builtin/fileformats/htmlformat.py
@@ -9,15 +9,15 @@ Basic implementation for a HTML importer
 """
 
 
-from mathics.builtin.base import Builtin
+from mathics.builtin.base import Builtin, MessageException
 from mathics.builtin.files_io.files import MathicsOpen
+from mathics.core.atoms import String
 from mathics.core.convert.expression import to_expression, to_mathics_list
+from mathics.core.convert.python import from_python
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
-from mathics.core.atoms import String, from_python
 from mathics.core.symbols import Symbol
 from mathics.core.systemsymbols import SymbolRule
-from mathics.builtin.base import MessageException
 
 from io import BytesIO
 import platform
