+++ mathics/core/evaluation.py
@@ -14,7 +14,8 @@ from mathics_scanner import TranslateError
 
 from mathics import settings
 
-from mathics.core.atoms import from_python, Integer, String
+from mathics.core.atoms import Integer, String
+from mathics.core.convert.python import from_python
 from mathics.core.element import KeyComparable, ensure_context
 from mathics.core.interrupt import (
     AbortInterrupt,
@@ -533,7 +534,7 @@ class Evaluation:
         self.output.out(self.out[-1])
 
     def print_out(self, text) -> None:
-        from mathics.core.atoms import from_python
+        from mathics.core.convert.python import from_python
 
         if self.definitions.trace_evaluation:
             self.definitions.trace_evaluation = False
